Done drafting SetNode vignette for now
Please read, review, add, and modify accordingly. I inserted a level-1
header in italics with a question for you (Benjamin).  It pertains to
how we best describe the factor conversions that take place in rToJags.

Maybe a short mention of this within the SettingNodes.Rmd vignette
would be helpful, with a more detailed explanation of going back and
forth before & after use of coda.samples given in the
ObtainPosteriorDensities.Rmd vignette?