tweaks to compileJagsModel()
I was getting an error message because the .factors object had stuff
that was not actually in the model. So I subsetted the .factors array
to only those variables actually in the HydeNetwork model.

I also tweaked the logic check for the cpt_arrays block.   The line

cpt_arrays <- unlist(network$nodeFitter) == "cpt"

gave a vector of length 4 equal to c(FALSE, FALSE, FALSE, FALSE).